AI_UseCase/requirements.txt
streamlit
pypdf
sentence-transformers
faiss-cpu
numpy
pandas
sqlite-utils
